Update sd-webUI.xml
Unraid support thread for Stable Diffusion WebUI

**CAUTION** The container downloads large diffusion models to model folder on run. (>4GB to docker folder) (>26GB download to model dir after creation)

This GPU-bound container requires the Nvidia driver installed on your Unraid server with support for Docker. This driver needs to support the version of CUDA in use by this container. The template adds --gpus all to the way the docker container is started to get access to the GPU(s).


The Unraid Nvidia Plugin is available in the community apps store

https://forums.unraid.net/topic/98978-plugin-nvidia-driver/

If you have multiple GPUs in your system with some allocated to VMs, make sure to replace --gpus all with --runtime=nvidia and follow the steps below to set the NVIDIA_DRIVER_CAPABILITIES and NVIDIA_VISIBLE_DEVICES variables to only give the container access to selected GPUs.
https://forums.unraid.net/topic/98978-plugin-nvidia-driver/?do=findComment&comment=913251